adjusting file deletion on S3

+++ app/lib/check_s3.py
@@ -24,7 +24,8 @@ from tornado.escape import json_decode
 from logging import info
 from datetime import datetime
 from json import dumps,loads
-from boto.s3.connection import S3Connection, Bucket, Key
+from boto import connect_s3
+from boto.s3.connection import Bucket, Key
 
 @gen.coroutine
 def checkS3(db,api):
@@ -37,7 +38,7 @@ def checkS3(db,api):
         S3_SECRET_KEY = api['S3_SECRET_KEY']
         S3_BUCKET = api['S3_BUCKET']
         try:
-            conn = S3Connection(S3_ACCESS_KEY,S3_SECRET_KEY)
+            conn = connect_s3(S3_ACCESS_KEY,S3_SECRET_KEY,is_secure=False,calling_format=OrdinaryCallingFormat())
             bucket = Bucket(conn, S3_BUCKET)
             info('\nConnected to S3')
         except:
@@ -48,11 +49,16 @@ def checkS3(db,api):
                 info('Removing files from: '+str(rmlist['ts']))
                 reqs = list()
                 try:
+                    alldeleted = True
                     for key in rmlist['list']:
                         info(str(key))
                         info(str(S3_BUCKET))
                         k = Key(bucket = bucket, name=key)
-                        k.delete()
-                    res = db.dellist.remove({'_id':rmlist['_id']})
+                        if k.exists():
+                            k.delete()
+                        if k.exists():
+                            alldeleted = False
+                    if alldeleted:
+                        res = db.dellist.remove({'_id':rmlist['_id']})
                 except:
                     info('Error in the deletion of images')
